comp1.1 - created binary and unary gate modules
